Dear Alex,
I found some issues with regards to the opar options which I fixed. I also change the title to RAPTOR: Position and row tracheid organizer in R as it is better English. I also have some issues with you change to plot_cells instead of plot.cells. I tried to change it back but in my version it does not work smoothly. Also, I think we need to check if there are issues with the write output function. Lets discuss this on sunday.

Cheers,

Richard