Dear Alex,
I changed mainly textual things and some of the coding. The test now works without any warnings in my version. My question for you is if you could check this version (focussing on the text) and potentially perform the final step to generate the package. Maybe we can skype on Tuesday to do this together?

All the best,

Richard